requirements.txt
aiomysql==0.0.21
aiosqlite==0.17.0
asgiref==3.4.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
idna==3.2
iso8601==0.1.16
isodate==0.6.0
lxml==4.6.3
platformdirs==2.4.0
pydantic==1.8.2
PyMySQL==0.9.3
pypika-tortoise==0.1.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
starlette==0.14.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
zeep==4.1.0
